requirements.txt
Flask
Flask-SQLAlchemy
Werkzeug
PyPDF2
requests
openai
python-dotenv
gunicorn
Flask-Migrate
azure-core==1.29.5
azure-identity==1.15.0

